pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-fashion-trends"
version = "0.1.0"
description = "Прогнозирование модных трендов с помощью ИИ-агентов"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pytrends>=4.9",
    "feedparser>=6.0",
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "statsmodels>=0.14",
    "numpy>=1.24",
    "pillow>=10.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
]

[project.optional-dependencies]
vision = [
    "mediapipe>=0.10.14",
]
torch = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
]

[project.scripts]
ai-fashion-trends = "ai_fashion_trends.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_fashion_trends"]
